feat : mysql -> postgresql로 마이그레이션

+++ src/main/kotlin/com/beyoureyes/beyoureyes/dto/FlatDailyFoodDto.kt
@@ -5,13 +5,13 @@ import java.time.LocalDateTime
 data class FlatDailyFoodDto(
     val logId: Int,
     val foodPhoto: String,
-    val calories: Double,
-    val carbohydrates: Double,
-    val protein: Double,
-    val fat: Double,
-    val cholesterol: Double,
-    val sodium: Double,
-    val sugar: Double,
-    val saturatedFat: Double,
+    val calories: Int,
+    val carbohydrates: Int,
+    val protein: Int,
+    val fat: Int,
+    val cholesterol: Int,
+    val sodium: Int,
+    val sugar: Int,
+    val saturatedFat: Int,
     val dateTime: LocalDateTime
 )
