+++ src/main/kotlin/com/beyoureyes/beyoureyes/dto/NutrientSummaryDto.kt
@@ -1,12 +1,12 @@
 package com.beyoureyes.beyoureyes.dto
 
 data class NutrientSummaryDto(
-    val calories: Double,
-    val carbohydrates: Double,
-    val protein: Double,
-    val fat: Double,
-    val cholesterol: Double,
-    val sodium: Double,
-    val sugar: Double,
-    val saturatedFat: Double
+    val calories: Int,
+    val carbohydrates: Int,
+    val protein: Int,
+    val fat: Int,
+    val cholesterol: Int,
+    val sodium: Int,
+    val sugar: Int,
+    val saturatedFat: Int
 )
